+++ src/engine_v0_2/src/Player.cpp
@@ -9,14 +9,14 @@
 
 using namespace ge;
 
-Player::Player( const std::vector<ge::Color>& colorPoll, const sf::Vector2f& initPosition, float baseSpeed , MovementController * movementController) : m_currentColorIndex(0), m_colorPoll(colorPoll), m_position(initPosition), m_baseSpeed(baseSpeed), m_movementController(movementController) {
+Player::Player( const std::vector<ge::Color>& colorPoll, const sf::Vector2f& initPosition, double baseSpeed , MovementController * movementController) : m_currentColorIndex(0), m_colorPoll(colorPoll), m_position(initPosition), m_baseSpeed(baseSpeed), m_movementController(movementController) {
 }
 
 Player::~Player() {
     delete m_shape;
 }
 
-void Player::update( float elapsed ){
+void Player::update( unsigned long long elapsed ){
     
     sf::Vector2f deplacement = m_movementController->move(elapsed);
     deplacement.x *= m_baseSpeed;
